+++ src/main/java/com/project/institute_management_software/dto/AddCourseRequest.java
@@ -14,6 +14,6 @@ public class AddCourseRequest {
     private String name;
     private String description;
     private String duration;
-    private int instructorId;
+    private Long instructorId;
 
 }
